add: frontend

+++ backend/controllers/product.controller.js
@@ -50,11 +50,15 @@ export const deleteProducts = async (req, res) => {
     const {id} = req.params
     // console.log("id: ", id);
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id" })
+    }
+
     try {
         await Product.findByIdAndDelete(id)
         res.status(201).json({ success: true, message: "Product deleted" })
     } catch (error) {
         console.log("Error in deleting products:", error.message);
-        res.status(404).json({ success: false, message: "Product not found" })
+        res.status(505).json({ success: false, message: "Server error" })
     }
 }
